pyproject.toml
[project]
name = "renfe-mcp"
version = "0.3.0"
description = "MCP server for Renfe train schedules and prices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "fastmcp>=0.7.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "json5>=0.12.0",
    "pydantic>=2.11.7",
    "httpx>=0.27.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/renfe_mcp"]

[project.scripts]
renfe-mcp = "renfe_mcp.server:mcp.run"
